% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoDiet.R
\docType{package}
\name{EcoDiet-package}
\alias{EcoDiet}
\alias{EcoDiet-package}
\title{EcoDiet: Creating a diet matrix from different data}
\description{
TO DO
    TO DO
}
\author{
\strong{Maintainer}: Heloise Thero \email{thero.heloise@gmail.com}

Authors:
\itemize{
  \item Pierre-Yves Hernvann
}

}
\keyword{internal}
