% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecodiet_example}
\alias{ecodiet_example}
\title{Toy example to test the EcoDiet package}
\format{A list of 8 elements:
\describe{
  \item{isotope_data}{a data frame with 3 variables: \code{Group}, 
  \code{d13C} and \code{d15N}.}

  \item{SCA_data}{A data frame with 10 variables, 
   each corresponding to the name of a trophic group:
   \code{Cod}, \code{Pout}, \code{Sardine}, \code{Shrimps},
   \code{Crabs}, \code{Bivalves}, \code{Worms}, \code{Zooplankton},
   \code{Phytoplankton} and \code{Detritus}.
   The number are the number of stomachs in which a specific prey 
   (the group on the row) was found inside a predator (the group on the column).}
   
  \item{nb_stom_SCA_data}{A data frame with 10 variables, 
   each corresponding to the name of a trophic group, 
   containing the number of full stomachs analyzed for each predator.}
}}
\usage{
ecodiet_example
}
\description{
This is a simulated toy dataset, made to illustrate the EcoDiet package functions.
}
\examples{
data("ecodiet_example")
}
\keyword{datasets}
