% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecodiet_example_isotope}
\alias{ecodiet_example_isotope}
\title{Artificial isotope data to test the EcoDiet package}
\format{A data frame with 3 variables: \code{Group}, 
  \code{d13C} and \code{d15N}}
\usage{
ecodiet_example_isotope
}
\description{
This is the isotopic data of a simulated toy dataset, 
made to illustrate the EcoDiet package functions.
}
\keyword{datasets}
