% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecodiet_example_stomach}
\alias{ecodiet_example_stomach}
\title{Artificial stomachal data to test the EcoDiet package}
\format{A data frame with 10 variables, 
   each corresponding to the name of a trophic group:
   \code{Cod}, \code{Pout}, \code{Sardine}, \code{Shrimps},
   \code{Crabs}, \code{Bivalves}, \code{Worms}, \code{Zooplankton},
   \code{Phytoplankton} and \code{Detritus}}
\usage{
ecodiet_example_stomach
}
\description{
This is the stomachal data of a simulated toy dataset, 
made to illustrate the EcoDiet package functions.
}
\keyword{datasets}
