% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_isotope_data.csv}
\alias{example_isotope_data.csv}
\title{Example isotope data}
\format{A table with 24 rows and 3 columns. Each row is an isotopic sample from one individual.
The columns are:
\describe{
  \item{group}{the trophic group the individual belonged to}
  \item{d13C}{the d13C measurement made on that individual}
  \item{d15N}{the d15N measurement made on that individual}
}}
\description{
This is a simulated isotopic dataset, made to illustrate the EcoDiet package functions.
}
