% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_stomach_data.csv}
\alias{example_stomach_data.csv}
\title{Example stomach data}
\format{A table with 5 rows and 5 columns. 
The first row contains the predators' names, the first column contains the preys' names.
Each cell contains the number of the predator's stomachs in which this prey was found.
The last row contains the total number of full stomachs for each predator.}
\description{
This is a simulated stomachal dataset, made to illustrate the EcoDiet package functions.
}
