% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_stomach_data}
\alias{example_stomach_data}
\title{Example stomach data}
\format{A table with 5 rows and 5 columns.
The headers contain the predators' names, the first column contains the preys' names.
Each cell contains the number of the predator's stomachs in which this prey was found.
The last row contains the total number of non-empty stomachs for each predator.}
\description{
This is an artificial and simple stomachal dataset, made to illustrate the package on a simple case.
All tables whose name start by "example" are describing different data from the same trophic groups.
}
