% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize the different variables of the EcoDiet model to start the MCMC chains}
\usage{
initialize_model(data, topo_run)
}
\arguments{
\item{data}{the list of the preprocessed data needed to compute the dimension of the variables}

\item{topo_run}{the topology matrix that is part of the raw data}
}
\value{
a list containing the initial values for each variable that need to be initialized
}
\description{
Initialize the different variables of the EcoDiet model to start the MCMC chains
}
\examples{
inits <- initialize_model(preprocessed_data, ecodiet_example$topo_run)

}
