% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the raw data and process it to feed the EcoDiet model}
\usage{
load_data(raw_data_list)
}
\arguments{
\item{raw_data_list}{the list containing the different raw data objects}
}
\value{
a list of processed data, ready to be run by the EcoDiet model
}
\description{
Load the raw data and process it to feed the EcoDiet model
}
\examples{
data <- load_data(ecodiet_example)

}
