% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_output_example}
\alias{mcmc_output_example}
\title{The MCMC output for running the example dataset}
\format{An object of class \code{matrix} with 3000 rows and 12 columns.}
\usage{
mcmc_output_example
}
\description{
This is the MCMC output for running the example dataset as illustrated in the introduction 
vignette (with 1e6 iterations, 1e3 adaptation steps) and with priors informed 
from the literature study.
This data is here so that the plot_results fonction can be illustrated on 
results that have converged.
}
\examples{
data("mcmc_output_example")
}
\keyword{datasets}
