% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data}
\alias{plot_data}
\title{Plot the raw data entered by the user.}
\usage{
plot_data(isotope_data = NULL, stomach_data = NULL)
}
\arguments{
\item{isotope_data}{the raw isotopic data}

\item{stomach_data}{the raw stomachal data if the user has it.
By default this variable is set on NULL and is not ploted.}
}
\description{
Plot the raw data entered by the user.
}
