% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_results}
\alias{plot_results}
\title{Plot the results of the EcoDiet model.}
\usage{
plot_results(mcmc_output)
}
\arguments{
\item{mcmc_output}{the mcmc.list object outputed by the run_model() function}
}
\description{
Plot the results of the EcoDiet model.
}
