% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{plot_test_ggplot2_figure}
\alias{plot_test_ggplot2_figure}
\title{Plot the mpg by wt of the mtcars data using the ggplot2 package.}
\usage{
plot_test_ggplot2_figure()
}
\description{
Plot the mpg by wt of the mtcars data using the ggplot2 package.
}
\examples{
plot_test_ggplot2_figure()
}
