% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data2.R
\name{preprocess_data2}
\alias{preprocess_data2}
\title{Load and preprocess the data to feed the EcoDiet model}
\usage{
preprocess_data2(
  isotope_data,
  trophic_enrichment_factor,
  literature_prior = FALSE,
  trophic_links = NULL,
  element_concentration = 1,
  stomach_data = NULL,
  literature_diets = NULL,
  nb_eta_prior = 0,
  nb_PI_prior = 0,
  nb_eta_to_PI = 5
)
}
\arguments{
\item{isotope_data}{the table containing the isotopic data in the specific format}

\item{trophic_enrichment_factor}{a vector containing the mean trophic enrichement factor 
corresponding to each column of the isotope data table (except the group column)}

\item{literature_prior}{a boolean (TRUE or FALSE) indicating whether the model will have 
prior distributions defined by a study of the literature}

\item{trophic_links}{a potential matrix entered by the user if she wants the model to 
investigate additionnal trophic links (by default it is NULL and defined from the stomach
data and the alpha priors if they are defined)}

\item{stomach_data}{the table containing the stomachal data in a specific format}

\item{literature_diets}{the diet proportions and their associated pedigrees found in the literature}

\item{nb_eta_prior}{the equivalent number of stomach for the literature priors on eta}

\item{nb_PI_prior}{the equivalent number of isotope for the literature priors on PI}

\item{nb_eta_to_PI}{the equivalent number of stomach for which eta will impact PI}
}
\value{
a list of preprocessed data, ready to be run by the EcoDiet model
}
\description{
Load and preprocess the data to feed the EcoDiet model
}
