% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{print_convergence_diagnostic}
\alias{print_convergence_diagnostic}
\title{Print a message if there is a convergence problem with the MCMC chains
This is an internal function called only in the run_model function below.}
\usage{
print_convergence_diagnostic(mcmc_output)
}
\arguments{
\item{mcmc_output}{the mcmc.list containing the variables to store, i.e., the pi and lambda variables}
}
\description{
Print a message if there is a convergence problem with the MCMC chains
This is an internal function called only in the run_model function below.
}
\keyword{internal}
