% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{realistic_stomach_data}
\alias{realistic_stomach_data}
\title{Realistic stomach data}
\format{
A table with 11 rows and 11 columns.
The headers contain the predators' names, the first column contains the preys' names.
Each cell contains the number of the predator's stomachs in which this prey was found.
The last row contains the total number of non-empty stomachs for each predator.
}
\description{
This is an artificial and realistic stomachal dataset, made to illustrate the package on a complex dataset.
All tables whose name start by "realistic" are describing different data from the same trophic groups.
}
