% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_stomach_data.R
\name{rescale_stomach_data}
\alias{rescale_stomach_data}
\title{Rescale the stomachal data}
\usage{
rescale_stomach_data(SCA_data, nb_stom_SCA_data)
}
\arguments{
\item{stomach_df}{the dataframe containing the stomachal data}

\item{isotope_data_path}{the dataframe containing the number of full stomachs for each predator}
}
\value{
the rescaled stomach dataframe
}
\description{
When the occurrences of preys in the stomach of predators are too small, 
it is often useful to rescale them. This function artificially increases the number of 
ocurrences in stomach for which a prey is found.
}
\examples{
rescaled_stomach <- rescale_stomach_data(ecodiet_example$SCA_data, ecodiet_example$nb_stom_SCA_data)

}
