% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model_rjags}
\alias{run_model_rjags}
\title{Run the MCMC chains for the EcoDiet model with the rjags package}
\usage{
run_model_rjags(data, inits, model_file, n_iter = 1000)
}
\description{
Run the MCMC chains for the EcoDiet model with the rjags package
}
