% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{show_mtcars}
\alias{show_mtcars}
\title{Plot the mpg by wt of the mtcars data using the base package}
\usage{
show_mtcars()
}
\description{
Plot the mpg by wt of the mtcars data using the base package
}
\examples{
show_mtcars()
}
