% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{show_mtcars2}
\alias{show_mtcars2}
\title{Plot the mpg by wt of the mtcars data using the ggplot2 package}
\usage{
show_mtcars2()
}
\description{
Plot the mpg by wt of the mtcars data using the ggplot2 package
}
\examples{
show_mtcars2()
}
