% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestEcoDiet.R
\docType{package}
\name{TestEcoDiet-package}
\alias{TestEcoDiet}
\alias{TestEcoDiet-package}
\title{TestEcoDiet: Creating a diet matrix from different data}
\description{
This is a test version of the EcoDiet package.
}
\author{
\strong{Maintainer}: Heloise Thero \email{thero.heloise@gmail.com}

Authors:
\itemize{
  \item Pierre-Yves Hernvann
}

}
\keyword{internal}
