% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Toy_example_data}
\alias{Toy_example_data}
\title{Toy example to test the EcoDiet package}
\format{A list of 8 elements}
\usage{
Toy_example_data
}
\description{
This is a simulated toy dataset, made to illustrate the EcoDiet package functions.
}
\examples{
data("Toy_example_data")
}
\keyword{datasets}
