% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoDiet.R
\docType{package}
\name{EcoDiet-package}
\alias{EcoDiet}
\alias{EcoDiet-package}
\title{EcoDiet: Estimating a Diet Matrix from Biotracer and Stomach Content Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Biotracers and stomach content analyses are combined in a Bayesian hierarchical model
    to estimate a probabilistic topology matrix (all trophic link probabilities) and a diet matrix 
    (all diet proportions).
    The package relies on the JAGS software and the 'rjags' package to run a Markov chain Monte Carlo 
    approximation of the different variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pyhernvann/EcoDiet}
  \item Report bugs at \url{https://github.com/pyhernvann/EcoDiet/issues}
}

}
\author{
\strong{Maintainer}: Pierre-Yves Hernvann \email{pierre.yves.hernvann@gmail.com}

Authors:
\itemize{
  \item Heloise Thero
  \item Didier Gascuel
  \item Dorothee Kopp
  \item Marianne Robert
  \item Jerome Guitton
  \item Etienne Rivot
}

}
\keyword{internal}
