% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoDiet.R
\docType{package}
\name{EcoDiet-package}
\alias{EcoDiet}
\alias{EcoDiet-package}
\title{EcoDiet: Estimating a diet matrix from isotopic and stomachal data sources}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The goal of the package is to estimate trophic links and a diet matrix 
    from isotopic and stomachal data using a Bayesian Hierarchical Model. The package 
    relies on rjags to run a MCMC approximation of the different variables.
}
\author{
\strong{Maintainer}: Heloise Thero \email{thero.heloise@gmail.com}

Authors:
\itemize{
  \item Pierre-Yves Hernvann
  \item Didier Gascuel
  \item Dorothee Kopp
  \item Marianne Robert
  \item Jerome Guitton
  \item Etienne Rivot
}

}
\keyword{internal}
