% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the stomachal and the isotopic data to feed the EcoDiet model.}
\usage{
load_data(stomach_data_path, isotope_data_path)
}
\arguments{
\item{stomach_data_path}{the path to access the stomachal data}

\item{isotope_data_path}{the path to access the isotopic data}
}
\value{
a list containing the stomachal and isotopic data
}
\description{
Load the stomachal and the isotopic data to feed the EcoDiet model.
}
\examples{
data <- load_data("data/example_stomach_data.csv", "data/example_isotope_data.csv")

}
