% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot the prior means or distributions for the eta and/or PI parameters}
\usage{
plot_prior(
  data,
  literature_prior,
  pred = NULL,
  prey = NULL,
  variable = c("eta", "PI")
)
}
\arguments{
\item{data}{the preprocessed data list outputed by the preprocess_data() function}

\item{literature_prior}{a boolean (TRUE or FALSE) indicating whether the model will have 
prior distributions defined by a study of the literature}

\item{pred}{the predator name for which we want to plot the probability densities}

\item{prey}{the prey(s) name for which we want to plot the probability densities}

\item{variable}{the variable(s) for which we want to plot the probability densities}
}
\description{
Plot the prior means or distributions for the eta and/or PI parameters
}
