% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_results}
\alias{plot_results}
\title{Plot the main results of the EcoDiet model (the means of each variable)}
\usage{
plot_results(
  mcmc_output,
  data,
  pred = NULL,
  prey = NULL,
  variables = c("PI", "eta")
)
}
\arguments{
\item{mcmc_output}{the mcmc.list object outputed by the run_model() function}

\item{data}{the preprocessed data outputed by the preprocess_data() function}
}
\description{
Plot the main results of the EcoDiet model (the means of each variable)
}
