% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Load and preprocess the data to feed the EcoDiet model}
\usage{
preprocess_data(
  stomach_data,
  isotope_data,
  trophic_enrichment_factor,
  literature_prior = FALSE,
  element_concentration = 1,
  trophic_links = NULL,
  literature_diets = NULL,
  literature_pedigrees = NULL,
  nb_literature = 10,
  literature_slope = 0.5
)
}
\arguments{
\item{stomach_data}{the table containing the stomachal data in a specific format}

\item{isotope_data}{the table containing the isotopic data in the specific format}

\item{trophic_enrichment_factor}{a vector containing the mean trophic enrichement factor 
corresponding to each column of the isotope data table (except the group column)}

\item{literature_prior}{a boolean (TRUE or FALSE) indicating whether the model will have 
prior distributions defined by a study of the literature}

\item{trophic_links}{a potential matrix entered by the user if she wants the model to 
investigate additionnal trophic links (by default it is NULL and defined from the stomach
data and the alpha priors if they are defined)}
}
\value{
a list of preprocessed data, ready to be run by the EcoDiet model
}
\description{
Load and preprocess the data to feed the EcoDiet model
}
