% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Load and preprocess the data to feed the EcoDiet model}
\usage{
preprocess_data(
  isotope_data,
  trophic_discrimination_factor,
  literature_configuration = FALSE,
  topology = NULL,
  element_concentration = 1,
  stomach_data = NULL,
  literature_diets = NULL,
  nb_literature,
  literature_slope
)
}
\arguments{
\item{isotope_data}{the table containing the isotopic data in the specific format}

\item{trophic_discrimination_factor}{a vector containing the trophic discrimination factors corresponding
to each biotracer found in the biotracer data
corresponding to each column of the isotope data table (except the group column)}

\item{literature_configuration}{a boolean (TRUE or FALSE) indicating whether the model will have
prior distributions informed by a literature study}

\item{topology}{a matrix that the user may input if she wants the model to
investigate some additionnal trophic links (by default it is NULL and defined from the stomach
data and the alpha priors if they are defined)}

\item{stomach_data}{the table containing the stomachal data in a specific format}

\item{literature_diets}{the diet proportions and their associated pedigrees found in the literature}

\item{nb_literature}{the equivalent number of stomach for the literature priors}

\item{literature_slope}{the slope of the linear relationship between the pedigrees
and the PIs' coefficients of variation (CVs)}
}
\value{
a list of preprocessed data, ready to be run by the EcoDiet model
}
\description{
Load and preprocess the data to feed the EcoDiet model
}
