% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Load and preprocess the raw data to feed the EcoDiet model}
\usage{
preprocess_data(raw_data_list)
}
\arguments{
\item{raw_data_list}{the list containing the different raw data objects}
}
\value{
a list of processed data, ready to be run by the EcoDiet model
}
\description{
Load and preprocess the raw data to feed the EcoDiet model
}
\examples{
data <- preprocess_data(ecodiet_example)

}
