% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the MCMC chains for the EcoDiet model}
\usage{
run_model(data, inits, model_file, n.iter = 1000)
}
\arguments{
\item{data}{the list containing the different loaded and preprocessed data objects}

\item{inits}{the list containing the initial values of the variables}

\item{model_string}{the string containing the definition of the EcoDiet model}
}
\value{
an mcmc.list containing the variables to store, i.e., the pi and lambda variables
}
\description{
Run the MCMC chains for the EcoDiet model
}
\examples{
mcmc_output <- run_model(preprocessed_data, model_string)

}
