% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the MCMC chains for the EcoDiet model with the rjags package}
\usage{
run_model(model_file, data, inits = NULL, n_iter = 100, n_chains = 3)
}
\arguments{
\item{model_file}{the file containing the definition of the EcoDiet model}

\item{data}{the list containing the different loaded and preprocessed data objects}

\item{inits}{the list containing the initial values of the variables.
By default the initialisation values are null, i.e., the first parameters at the beginning
of the chains are drawns randomly from their prior distribution.}

\item{n_iter}{the number of iterations to be run}

\item{n_chains}{the number of Markov chains to be run}
}
\value{
an mcmc.list containing the variables to store, i.e., the pi and lambda variables
}
\description{
Run the MCMC chains for the EcoDiet model with the rjags package
}
\examples{
mcmc_output <- run_model(preprocessed_data, inits, model_string)

}
