% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write a string containing the EcoDiet model in the BUGS syntax}
\usage{
write_model(literature_prior = FALSE)
}
\arguments{
\item{file_name}{the name of the bugs files on which the model will be written}
}
\value{
a string containing the model definition written in bugs syntax
}
\description{
Write a string containing the EcoDiet model in the BUGS syntax
}
\examples{
model_string <- write_model()

}
