% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write a string containing the EcoDiet model in the BUGS syntax}
\usage{
write_model(literature_prior = FALSE)
}
\arguments{
\item{literature_prior}{a boolean (TRUE or FALSE) indicating whether the model will have 
prior distributions defined by a study of the literature}
}
\value{
a string containing the model definition written in BUGS syntax
}
\description{
Write a string containing the EcoDiet model in the BUGS syntax
}
\examples{
model_string <- write_model()

}
