% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Write the EcoDiet model in bugs langage on a file}
\usage{
write_model()
}
\arguments{
\item{file_name}{the name of the bugs files on which the model will be written}
}
\value{
a string containing the model definition written in bugs syntaw
}
\description{
Write the EcoDiet model in bugs langage on a file
}
\examples{
model_def <- write_model()

}
